backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
xgboost==2.0.2
prophet==1.1.4
statsmodels==0.14.0

# Data Processing
numpy==1.24.3
pandas==2.1.4
geopandas==0.14.1
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Weather Data
# NOAA data collection (free, no API keys needed)

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching and Background Tasks
redis==5.0.1
celery==5.3.4

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
